JW-1057: divide driver functions into separate files
